requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.3.0
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
multidict==5.0.2
oauthlib==3.1.0
pycparser==2.20
python-dateutil==2.8.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
